separate files for hybrid and continuous torus
